+++ src/main/java/hello/TypeCustomerRepository.java
@@ -2,5 +2,7 @@ package hello;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-interface TypeCustomerRepository extends JpaRepository<TypeCustomer, Long> {
+import java.math.BigDecimal;
+
+interface TypeCustomerRepository extends JpaRepository<TypeCustomer, BigDecimal> {
 }
